// generated/dto/update-project-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Auth API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateProjectDto
 */
export interface UpdateProjectDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateProjectDto
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateProjectDto
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateProjectDto
     */
    'custom_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateProjectDto
     */
    'lead_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateProjectDto
     */
    'cover_image'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateProjectDto
     */
    'logo'?: string;
}